+++ examples/resources/harness_platform_cluster/import.sh
@@ -1,2 +1,0 @@
-# Import using the cluster.
-terraform import harness_platform_cluster.example <cluster_id>