+++ examples/resources/harness_platform_environment/resource.tf
@@ -1,0 +1,53 @@
+resource "harness_platform_environment" "test" {
+  identifier = "identifier"
+  name       = "name"
+  org_id     = "org_id"
+  project_id = "project_id"
+  tags       = ["foo:bar", "baz"]
+  type       = "PreProduction"
+  yaml       = <<-EOT
+			   environment:
+         name: name
+         identifier: identifier
+         orgIdentifier: org_id
+         projectIdentifier: project_id
+         type: PreProduction
+         tags:
+           foo: bar
+           baz: ""
+         variables:
+           - name: envVar1
+             type: String
+             value: v1
+             description: ""
+           - name: envVar2
+             type: String
+             value: v2
+             description: ""
+         overrides:
+           manifests:
+             - manifest:
+                 identifier: manifestEnv
+                 type: Values
+                 spec:
+                   store:
+                     type: Git
+                     spec:
+                       connectorRef: <+input>
+                       gitFetchType: Branch
+                       paths:
+                         - file1
+                       repoName: <+input>
+                       branch: master
+           configFiles:
+             - configFile:
+                 identifier: configFileEnv
+                 spec:
+                   store:
+                     type: Harness
+                     spec:
+                       files:
+                         - account:/Add-ons/svcOverrideTest
+                       secretFiles: []
+      EOT
+}