+++ examples/resources/harness_platform_environment_clusters_mapping/import.sh
@@ -1,0 +1,2 @@
+# Import using the cluster.
+terraform import harness_platform_environment_clusters_mapping.example <cluster_id>