+++ helpers/errors.go
@@ -9,10 +9,14 @@ import (
 )
 
 func HandleApiError(err error, d *schema.ResourceData, httpResp *http.Response) diag.Diagnostics {
-	if httpResp != nil && httpResp.StatusCode == 401 {
-		return diag.Errorf(httpResp.Status + "\n" + "Hint:\n" +
-			"1) Please check if token has expired or is wrong.\n" +
-			"2) Harness Provider is misconfigured. For firstgen resources please give the correct api_key and for nextgen resources please give the correct platform_api_key.")
+	erro, ok := err.(nextgen.GenericSwaggerError)
+	if ok {
+		if httpResp != nil && httpResp.StatusCode == 401 {
+			return diag.Errorf(httpResp.Status + "\n" + "Hint:\n" +
+				"1) Please check if token has expired or is wrong.\n" +
+				"2) Harness Provider is misconfigured. For firstgen resources please give the correct api_key and for nextgen resources please give the correct platform_api_key.")
+		}
+		return diag.Errorf(erro.Error())
 	}
-	return diag.Errorf(err.(nextgen.GenericSwaggerError).Error())
+	return diag.Errorf(err.Error())
 }