+++ internal/service/platform/environment_clusters_mapping/data_source_environment_clusters_mapping.go
@@ -11,11 +11,11 @@ import (
 	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/schema"
 )
 
-func DataSourceCluster() *schema.Resource {
+func DataSourceEnvironmentClustersMapping() *schema.Resource {
 	resource := &schema.Resource{
 		Description: "Data source for retrieving a Harness Cluster.",
 
-		ReadContext: dataSourceClusterRead,
+		ReadContext: dataSourceResourceEnvironmentClustersMappingRead,
 
 		Schema: map[string]*schema.Schema{
 			"identifier": {
@@ -38,16 +38,21 @@ func DataSourceCluster() *schema.Resource {
 				Type:        schema.TypeString,
 				Optional:    true,
 			},
+			"scope": {
+				Description: "scope at which the cluster exists in harness gitops",
+				Type:        schema.TypeString,
+				Computed:    true,
+			},
 		},
 	}
 	return resource
 }
 
-func dataSourceClusterRead(ctx context.Context, d *schema.ResourceData, meta interface{}) diag.Diagnostics {
+func dataSourceResourceEnvironmentClustersMappingRead(ctx context.Context, d *schema.ResourceData, meta interface{}) diag.Diagnostics {
 	c, ctx := meta.(*internal.Session).GetPlatformClientWithContext(ctx)
 	id := d.Get("identifier").(string)
 
-	resp, httpResp, err := c.ClustersApi.GetCluster(ctx, id, c.AccountId, id, &nextgen.ClustersApiGetClusterOpts{
+	resp, httpResp, err := c.ClustersApi.GetCluster(ctx, id, c.AccountId, d.Get("env_id").(string), &nextgen.ClustersApiGetClusterOpts{
 		OrgIdentifier:     optional.NewString(d.Get("org_id").(string)),
 		ProjectIdentifier: optional.NewString(d.Get("project_id").(string)),
 	})
@@ -64,9 +69,7 @@ func dataSourceClusterRead(ctx context.Context, d *schema.ResourceData, meta int
 		return nil
 	}
 
-	readCluster(d, resp.Data)
+	readEnvironmentClustersMappingCluster(d, resp.Data)
 
 	return nil
 }
-
-