+++ internal/service/platform/service/resource_service_test.go
@@ -49,6 +49,44 @@ func TestAccResourceService(t *testing.T) {
 	})
 }
 
+func TestAccResourceServiceWithYaml(t *testing.T) {
+
+	name := t.Name()
+	id := fmt.Sprintf("%s_%s", name, utils.RandStringBytes(5))
+	updatedName := fmt.Sprintf("%s_updated", name)
+	varValue := t.Name()
+	updatedVarValue := fmt.Sprintf("%s_updated", varValue)
+	resourceName := "harness_platform_service.test"
+
+	resource.UnitTest(t, resource.TestCase{
+		PreCheck:          func() { acctest.TestAccPreCheck(t) },
+		ProviderFactories: acctest.ProviderFactories,
+		CheckDestroy:      testAccServiceDestroy(resourceName),
+		Steps: []resource.TestStep{
+			{
+				Config: testAccResourceServiceWithYaml(id, name, varValue),
+				Check: resource.ComposeTestCheckFunc(
+					resource.TestCheckResourceAttr(resourceName, "id", id),
+					resource.TestCheckResourceAttr(resourceName, "name", name),
+				),
+			},
+			{
+				Config: testAccResourceServiceWithYaml(id, updatedName, updatedVarValue),
+				Check: resource.ComposeTestCheckFunc(
+					resource.TestCheckResourceAttr(resourceName, "id", id),
+					resource.TestCheckResourceAttr(resourceName, "name", updatedName),
+				),
+			},
+			{
+				ResourceName:      resourceName,
+				ImportState:       true,
+				ImportStateVerify: true,
+				ImportStateIdFunc: acctest.ProjectResourceImportStateIdFunc(resourceName),
+			},
+		},
+	})
+}
+
 func TestAccResourceService_DeleteUnderlyingResource(t *testing.T) {
 	t.Skip()
 	name := t.Name()
@@ -84,14 +122,14 @@ func TestAccResourceService_DeleteUnderlyingResource(t *testing.T) {
 	})
 }
 
-func testAccGetService(resourceName string, state *terraform.State) (*nextgen.EnvironmentResponseDetails, error) {
+func testAccGetService(resourceName string, state *terraform.State) (*nextgen.ServiceResponseDetails, error) {
 	r := acctest.TestAccGetResource(resourceName, state)
 	c, ctx := acctest.TestAccGetPlatformClientWithContext()
 	id := r.Primary.ID
 	orgId := r.Primary.Attributes["org_id"]
 	projId := r.Primary.Attributes["project_id"]
 
-	resp, _, err := c.EnvironmentsApi.GetEnvironmentV2(ctx, id, c.AccountId, &nextgen.EnvironmentsApiGetEnvironmentV2Opts{
+	resp, _, err := c.ServicesApi.GetServiceV2(ctx, id, c.AccountId, &nextgen.ServicesApiGetServiceV2Opts{
 		OrgIdentifier:     optional.NewString(orgId),
 		ProjectIdentifier: optional.NewString(projId),
 	})
@@ -100,11 +138,11 @@ func testAccGetService(resourceName string, state *terraform.State) (*nextgen.En
 		return nil, err
 	}
 
-	if resp.Data == nil || resp.Data.Environment == nil {
+	if resp.Data == nil || resp.Data.Service == nil {
 		return nil, nil
 	}
 
-	return resp.Data.Environment, nil
+	return resp.Data.Service, nil
 }
 
 func testAccServiceDestroy(resourceName string) resource.TestCheckFunc {
@@ -120,23 +158,86 @@ func testAccServiceDestroy(resourceName string) resource.TestCheckFunc {
 
 func testAccResourceService(id string, name string) string {
 	return fmt.Sprintf(`
-		resource "harness_platform_organization" "test" {
-			identifier = "%[1]s"
-			name = "%[2]s"
-		}
+    resource "harness_platform_organization" "test" {
+      identifier = "%[1]s"
+      name = "%[2]s"
+    }
 
-		resource "harness_platform_project" "test" {
-			identifier = "%[1]s"
-			name = "%[2]s"
-			org_id = harness_platform_organization.test.id
-			color = "#472848"
-		}
+    resource "harness_platform_project" "test" {
+      identifier = "%[1]s"
+      name = "%[2]s"
+      org_id = harness_platform_organization.test.id
+      color = "#472848"
+  }
 
-		resource "harness_platform_service" "test" {
-			identifier = "%[1]s"
-			name = "%[2]s"
-			org_id = harness_platform_project.test.org_id
-			project_id = harness_platform_project.test.id
-		}
+    resource "harness_platform_service" "test" {
+      identifier = "%[1]s"
+      name = "%[2]s"
+      org_id = harness_platform_project.test.org_id
+      project_id = harness_platform_project.test.id
+    }
 `, id, name)
+}
+
+func testAccResourceServiceWithYaml(id string, name string, varValue string) string {
+	return fmt.Sprintf(`
+    resource "harness_platform_organization" "test" {
+      identifier = "%[1]s"
+      name = "%[2]s"
+    }
+
+    resource "harness_platform_project" "test" {
+      identifier = "%[1]s"
+      name = "%[2]s"
+      org_id = harness_platform_organization.test.id
+      color = "#472848"
+    }
+
+    resource "harness_platform_service" "test" {
+      identifier = "%[1]s"
+      name = "%[2]s"
+      org_id = harness_platform_project.test.org_id
+      project_id = harness_platform_project.test.id
+      yaml = <<-EOT
+        service:
+          name: %[1]s
+          identifier: %[2]s
+          serviceDefinition:
+            spec:
+              manifests:
+                - manifest:
+                    identifier: manifest1
+                    type: K8sManifest
+                    spec:
+                      store:
+                        type: Github
+                        spec:
+                          connectorRef: <+input>
+                          gitFetchType: Branch
+                          paths:
+                            - files1
+                          repoName: <+input>
+                          branch: master
+                      skipResourceVersioning: false
+              configFiles:
+                - configFile:
+                    identifier: configFile1
+                    spec:
+                      store:
+                        type: Harness
+                        spec:
+                          files:
+                            - <+org.description>
+              variables:
+                - name: var1
+                  type: String
+                  value: %[3]s
+                - name: var2
+                  type: String
+                  value: val2
+            type: Kubernetes
+          gitOpsEnabled: false
+      EOT
+    }
+`, id, name, varValue)
 }