+++ internal/service/platform/service_account/service_account.go
@@ -24,12 +24,12 @@ func ResourceServiceAccount() *schema.Resource {
 			"email": {
 				Description: "Email of the Service Account.",
 				Type:        schema.TypeString,
-				Optional:    true,
+				Required:    true,
 			},
 			"account_id": {
 				Description: "Account Identifier for the Entity.",
 				Type:        schema.TypeString,
-				Optional:    true,
+				Required:    true,
 			},
 		},
 	}